Design grid views to display coin details

+++ CryptoTrackerBootcamp/CryptoTrackerBootcamp/Models/CoinModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct CoinModel: Identifiable, Codable, Hashable {
     let id, symbol, name, image: String
     let currentPrice: Double
-    let marketCap, marketCapRank, fullyDilutedValuation: Int?
+    let marketCap, marketCapRank, fullyDilutedValuation: Double?
     let totalVolume, high24H, low24H: Double?
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H, marketCapChangePercentage24H: Double?
